refactor: Player 패키지 변경
