+++ src/main/kotlin/com/example/kbocombo/crawler/player/application/PlayerClient.kt
@@ -1,6 +1,6 @@
 package com.example.kbocombo.crawler.player.application
 
-import com.example.kbocombo.player.Player
+import com.example.kbocombo.player.domain.Player
 
 interface PlayerClient {
